make virtualenv shared to host for anaconda-mode.
Rather than pip install to the system inside docker, use a virtualenv.
The Dockerfile builds up the virtualenv in the image for deployment
purposes, but that virtualenv isn't visible to the host. So for
development, make the virtualenv and pip install from the Makefile as
well, so that it lives on the host system. This also makes it easy to
update for new requirements by hitting ctrl-c and rerunning "make dev"